backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.4.0
python-multipart>=0.0.6
sendgrid>=6.10.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.108.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dateutil>=2.8.0 